Moved native libs into native directory, added dynamic exports based on node|browser
